Update 02_network
First, thanks for the Rockchip images for Openwrt!

On the NanoPi R6S the ports are incorrectly mapped. This is the current situation (see screenshot).
The ports are currently mapped one to the right.

The correct mapping that represents the physical ports should be:
eth2 to eth1
eth1 to eth0
eth0 to eth2.